+++ hw/code/ui/locators/ad_creation_page_locators.py
